Fixed a bug when measuring long-running code.

+++ Launchers/RunP2PAlgo.cc
@@ -1,4 +1,5 @@
 #include <chrono>
+#include <cstdint>
 #include <cstdlib>
 #include <fstream>
 #include <iostream>
@@ -77,12 +78,12 @@ inline void writeHeaderLine(std::ofstream& out, AlgoT&) {
 
 // Writes a record line of the output CSV file, containing statistics about a single query.
 template <typename AlgoT>
-inline void writeRecordLine(std::ofstream& out, AlgoT& algo, const int, const int elapsed) {
+inline void writeRecordLine(std::ofstream& out, AlgoT& algo, const int, const int64_t elapsed) {
   out << algo.getDistance() << ',' << elapsed << '\n';
 }
 
 template <>
-inline void writeRecordLine(std::ofstream& out, Dij& algo, const int dst, const int elapsed) {
+inline void writeRecordLine(std::ofstream& out, Dij& algo, const int dst, const int64_t elapsed) {
   out << algo.getDistance(dst) << ',' << elapsed << '\n';
 }
